# apps/server/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Hickory Docs API",
    "description": "",
    "license": {
      "name": "GPL-3.0-or-later",
      "identifier": "GPL-3.0-or-later"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/billing/checkout": {
      "post": {
        "tags": [
          "billing"
        ],
        "operationId": "checkout",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckoutRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Stripe Checkout session url",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckoutOut"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/docs/{id}/outputs": {
      "get": {
        "tags": [
          "outputs"
        ],
        "summary": "GET /api/docs/:id/outputs → `{files: [{path, language}]}`",
        "operationId": "list_outputs",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "doc id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "files from the last successful run",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OutputsListOut"
                }
              }
            }
          }
        }
      }
    },
    "/api/docs/{id}/outputs/edit": {
      "post": {
        "tags": [
          "outputs"
        ],
        "summary": "POST /api/docs/:id/outputs/edit → `{source_edits, applied: true}`",
        "operationId": "edit_outputs",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "doc id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EditRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "applied source edits",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EditOutputsOut"
                }
              }
            }
          }
        }
      }
    },
    "/api/docs/{id}/outputs/file": {
      "get": {
        "tags": [
          "outputs"
        ],
        "summary": "GET /api/docs/:id/outputs/file?path=<rel> →\n`{path, language, content, provenance: Provenance[]}`",
        "operationId": "get_output_file",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "doc id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "path",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "output file content + provenance",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OutputFileOut"
                }
              }
            }
          }
        }
      }
    },
    "/api/docs/{id}/outputs/nav": {
      "post": {
        "tags": [
          "outputs"
        ],
        "summary": "`{path, offset, kind: \"definition\"|\"references\"}` →\n`{targets: [{uri, range: {start, end}}]}` (byte offsets).",
        "description": "The output byte offset is mapped through the last run's provenance into\nsource-document coordinates, a short-lived hick-lsp session answers the\nrequest over the seeded checkout, and results are mapped back: locations\nin `.hick` docs become `hick:///<doc-path>` byte ranges; locations that\nstay inside generated outputs are re-mapped through provenance where\npossible, else returned as `hick-output:///<output-path>` byte ranges.\nRead access follows doc visibility, like every other doc read.",
        "operationId": "outputs_nav",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "doc id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NavRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "definition/reference targets",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NavOut"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CheckoutOut": {
        "type": "object",
        "required": [
          "checkout_url"
        ],
        "properties": {
          "checkout_url": {
            "type": "string"
          }
        }
      },
      "CheckoutRequest": {
        "type": "object",
        "required": [
          "price_key"
        ],
        "properties": {
          "price_key": {
            "type": "string"
          }
        }
      },
...
      "EditOutputsOut": {
        "type": "object",
        "description": "`source_edits` is `hickory_lineage::SourceEdit[]` — same external-type\nrationale as `EditRequest::edits` above.",
        "required": [
          "source_edits",
          "applied"
        ],
        "properties": {
          "applied": {
            "type": "boolean"
          },
          "source_edits": {
            "type": "array",
            "items": {
              "type": "object"
            }
          }
        }
      },
      "EditRequest": {
        "type": "object",
        "required": [
          "path",
          "edits"
        ],
        "properties": {
          "edits": {
            "type": "array",
            "items": {
              "type": "object"
            },
            "description": "`hickory_lineage::OutputEdit` — an external crate type with no\n`ToSchema` impl, so it's documented as a generic JSON array rather\nthan pulling utoipa into `hickory-lineage`."
          },
          "path": {
            "type": "string"
          }
        }
      },
...
      "NavOut": {
        "type": "object",
        "description": "`targets` items are `{uri, range: {start, end}}`, built ad hoc from the\nLSP response's `Location | Location[] | LocationLink[]` union below —\nkept generic rather than asserting a schema over an upstream union this\nhandler only partially normalizes.",
        "required": [
          "targets"
        ],
        "properties": {
          "targets": {
            "type": "array",
            "items": {
              "type": "object"
            }
          }
        }
      },
      "NavRequest": {
        "type": "object",
        "required": [
          "path",
          "offset",
          "kind"
        ],
        "properties": {
          "kind": {
            "type": "string"
          },
          "offset": {
            "type": "integer",
            "minimum": 0
          },
          "path": {
            "type": "string"
          }
        }
      },
      "OutputFileMeta": {
        "type": "object",
        "required": [
          "path",
          "language"
        ],
        "properties": {
          "language": {
            "type": "string"
          },
          "path": {
            "type": "string"
          }
        }
      },
      "OutputFileOut": {
        "type": "object",
        "description": "`provenance` is `hickory_lineage::Provenance[]` — an external crate type\nwith no `ToSchema` impl, kept generic rather than pulling utoipa into\n`hickory-lineage`.",
        "required": [
          "path",
          "language",
          "content",
          "provenance"
        ],
        "properties": {
          "content": {
            "type": "string"
          },
          "language": {
            "type": "string"
          },
          "path": {
            "type": "string"
          },
          "provenance": {
            "type": "array",
            "items": {
              "type": "object"
            }
          }
        }
      },
      "OutputsListOut": {
        "type": "object",
        "required": [
          "files"
        ],
        "properties": {
          "files": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OutputFileMeta"
            }
          }
        }
      },
...
    }
  }
}